requirements.txt
pandas
torch>=1.7.0
numpy
PyYAML
scikit_learn
torch_optimizer